# Script to trigger build/deploy to dev and other refactors and improvements