--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -33,6 +33,9 @@ commands:
     parameters:
       study_key:
         type: string
+      study_guid:
+        type: string
+        default: NOTNEEDED
       skip_build:
         type: boolean
         default: false