--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -402,14 +402,11 @@ jobs:
           study_key: << pipeline.parameters.study_key >>
 
   deploy-stored-build-job:
-    parameters:
-      study_key:
-        type: string
     executor:
       name: commit-executor
     steps:
       - deploy-stored-build:
-          study_key: <<parameter.study_key>>
+          study_key: << pipeline.parameters.study_key >>
 
   osteo-deploy-only:
     executor:
@@ -525,13 +522,13 @@ workflows:
                 - production
           study_key: osteo
 
-      - deploy-stored-build:
+      - deploy-stored-build-job:
           <<: *deploy_branch_filters
           study_key: brain
-      - deploy-stored-build:
+      - deploy-stored-build-job:
           <<: *deploy_branch_filters
           study_key: mbc
-      - deploy-stored-build:
+      - deploy-stored-build-job:
           <<: *deploy_branch_filters
           study_key: angio