--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -402,11 +402,14 @@ jobs:
           study_key: << pipeline.parameters.study_key >>
 
   deploy-stored-build-job:
+    parameters:
+      study_key:
+        type: string
     executor:
       name: commit-executor
     steps:
       - deploy-stored-build:
-          study_key: << pipeline.parameters.study_key >>
+          study_key: << parameters.study_key >>
 
   osteo-deploy-only:
     executor: