--- a/build-utils/Dockerfile-BuildDeploy
+++ b/build-utils/Dockerfile-BuildDeploy
@@ -111,7 +111,7 @@ RUN apk update
 RUN apk upgrade
 RUN apk add curl bash jq
 # Install ruby and ruby-bundler
-RUN apk add ruby ruby-bundler
+RUN apk add ruby ruby-bundler ruby-json
 
 # Setting up Angular CLI globally
 ENV ANGULAR_CLI @angular/cli@8.3.23