--- a/build-utils/Dockerfile-BuildDeploy
+++ b/build-utils/Dockerfile-BuildDeploy
@@ -114,7 +114,7 @@ RUN apk add --no-cache --virtual .build-deps \
     ruby \
     ruby-bundler \
     ruby-json \
-&& apk del. build-deps
+&& apk del .build-deps
 
 # Setting up Angular CLI globally
 ENV ANGULAR_CLI @angular/cli@8.3.23