Added script for situations where we need to deploy to dev from own branch and want to avoid hacking config.yml script further.
**Updates on April 9, 2020**
- Some refactoring of the CircleCi configuration. Objective was to remove need to keep passing around study guids (we should know what they are) and to remove repetitive volume.
- Script for remote build deploys no longer needs the study guid to be specified.
- Updated Docker file to be based on Alpine . It is __**MUCH**__ faster! Loading the image from CircleCi went down from 50 to 80 seconds down to 10 to 15 sec.
Image size went from 760MB to 160MB. Whoa!

Comments added throughout.