@fat23cat 2020-04-09
Am I right that with this script we just need to add new `study_key`, `study_guid` and executors in circleci config without specifying local branch to deploy new app into GAE?