@MocanaAtBroad 2020-04-09
> Am I right that with this script we just need to add new `study_key`, `study_guid` and executors in circleci config without specifying local branch to deploy new app into GAE?

No. You execute the script locally/from your computer. Will need to generate a CircleCi token as described, but that is it.