@fat23cat 2020-04-09
@MocanaAtBroad, could you provide example of how to run the script?